backend/requirements.txt
Flask==2.0.3
psycopg2==2.9.3
google-cloud-storage==1.43.0
python-dotenv==0.19.2
Werkzeug==2.0.3
